Create tip calculator
/*This code receives the value of the user's bill (the value of a restaurant bill), asks if he wants to pay the service charge and if he wants to offer a tip.
If he wants to pay the fee, it will be added to the total amount. If he wants to pay only the tip, a text will appear on the screen asking for the amount of the tip, and then adding the amount to the total amount.
If he wanted to pay both, both will happen. If he didn't want to pay anything, the total amount will be the amount of consumption without any additions.
The code is in Portuguese*/